[fix CD登録・曲登録]

+++ app/controllers/admin/musics_controller.rb
@@ -1,28 +1,33 @@
 class Admin::MusicsController < ApplicationController
   def new
-    @musics = Music.new
+    if current_user.admin != true
+      redirect_to public_cds_path
+    end
+    @music = Music.new
     @cds = Cd.where(deleated: 'false')
   end
 
   def create
-    @musics = Music.new(musics_params)
-    @musics.save
-    redirect_to public_cd_path(@music.cd_id)
+    musics = musics_params
+    musics.each do |music|
+      m = Music.new(music_params(music))
+      m.save
+    end
+    redirect_to public_cds_path
   end
 
   def edit
   end
 
   def update
-    @musics = musics_params.keys.each do |id|
-      music = Music.find(id)
-      music.update_attributes(music_params[id])
-      music
-    end
-    respond_with(@musics, location: edit_admin_music_path)
+    @music = Music.find(params[:id])
+    @music.update(music_params)
   end
 
   def destroy
+    if current_user.admin != true
+      redirect_to public_cds_path
+    end
     music = Music.find(params[:id])
     music.destroy
     redirect_to public_cd_path(music.cd_id)
@@ -30,6 +35,10 @@ class Admin::MusicsController < ApplicationController
 
   private
     def musics_params
-      params.require(:music).permit(:id, :disk_number, :track, :name, :cd_id])
+        params.require(:musics)
+    end
+
+    def music_params(music)
+      music.require(:music).permit(:name, :cd_id, :disk_number, :track)
     end
 end
